requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
babel==2.10.3
    # via cubicweb-francearchives (setup.py)
beautifulsoup4==4.11.1
    # via cubicweb-elasticsearch
boto3==1.24.27
    # via cubicweb-s3storage
botocore==1.27.27
    # via
    #   boto3
    #   s3transfer
cachetools==5.2.0
    # via cubicweb
certifi==2022.6.15
    # via
    #   elasticsearch
    #   requests
    #   sentry-sdk
charset-normalizer==2.1.0
    # via requests
cubicweb==3.37.8
    # via
    #   cubicweb-addressbook
    #   cubicweb-card
    #   cubicweb-compound
    #   cubicweb-eac
    #   cubicweb-elasticsearch
    #   cubicweb-file
    #   cubicweb-francearchives (setup.py)
    #   cubicweb-geocoding
    #   cubicweb-link
    #   cubicweb-oaipmh
    #   cubicweb-prometheus
    #   cubicweb-prov
    #   cubicweb-s3storage
    #   cubicweb-sentry
    #   cubicweb-skos
    #   cubicweb-varnish
cubicweb-addressbook==1.12.0
    # via cubicweb-eac
cubicweb-card==1.3.0
    # via cubicweb-francearchives (setup.py)
cubicweb-compound==0.9.0
    # via cubicweb-eac
cubicweb-eac==1.5.2
    # via cubicweb-francearchives (setup.py)
cubicweb-elasticsearch==0.13.0
    # via cubicweb-francearchives (setup.py)
cubicweb-file==3.4.0
    # via cubicweb-francearchives (setup.py)
cubicweb-geocoding==0.5.0
    # via cubicweb-addressbook
cubicweb-link==1.11.1
    # via cubicweb-francearchives (setup.py)
cubicweb-oaipmh==0.9.0
    # via cubicweb-francearchives (setup.py)
cubicweb-prometheus==0.3.0
    # via cubicweb-francearchives (setup.py)
cubicweb-prov==0.6.0
    # via
    #   cubicweb-eac
    #   cubicweb-francearchives (setup.py)
cubicweb-s3storage==3.2.0
    # via cubicweb-francearchives (setup.py)
cubicweb-sentry==0.7.0
    # via cubicweb-francearchives (setup.py)
cubicweb-skos==2.3.0
    # via
    #   cubicweb-eac
    #   cubicweb-francearchives (setup.py)
cubicweb-varnish==0.7.0
    # via cubicweb-francearchives (setup.py)
cwtags==1.2.3
    # via
    #   cubicweb-elasticsearch
    #   cubicweb-francearchives (setup.py)
deprecated==1.2.13
    # via redis
docutils==0.19.0
    # via cubicweb-card
elasticsearch==7.12.0
    # via
    #   cubicweb-elasticsearch
    #   elasticsearch-dsl
elasticsearch-dsl==7.3.0
    # via cubicweb-elasticsearch
filelock==3.7.1
    # via cubicweb
geojson==2.5.0
    # via cubicweb-francearchives (setup.py)
glamconv==0.2.0
    # via cubicweb-francearchives (setup.py)
html2text==2020.1.16
    # via logilab-mtconverter
hupper==1.10.3
    # via pyramid
idna==3.3
    # via requests
importlib-metadata==4.12.0
    # via markdown
isodate==0.6.1
    # via
    #   cubicweb-oaipmh
    #   rdflib
jinja2==3.1.2
    # via cubicweb-francearchives (setup.py)
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
logilab-common==1.9.7
    # via
    #   cubicweb
    #   cwtags
    #   logilab-constraint
    #   logilab-database
    #   logilab-mtconverter
    #   rql
    #   yams
logilab-constraint==0.6.2
    # via rql
logilab-database==1.18.1
    # via
    #   cubicweb
    #   rql
    #   yams
logilab-mtconverter==0.10.0
    # via cubicweb
lxml==4.9.1
    # via
    #   cubicweb
    #   cubicweb-oaipmh
    #   glamconv
    #   logilab-mtconverter
    #   sickle
markdown==3.4.1
    # via
    #   cubicweb
    #   markdown-urlize
markdown-urlize==0.2.0
    # via cubicweb
markupsafe==2.1.1
    # via jinja2
mypy-extensions==0.4.3
    # via logilab-common
packaging==21.3
    # via redis
passlib==1.7.4
    # via cubicweb
pastedeploy==2.1.1
    # via plaster-pastedeploy
pillow==9.2.0
    # via
    #   cubicweb-file
    #   cubicweb-francearchives (setup.py)
plaster==1.0
    # via
    #   plaster-pastedeploy
    #   pyramid
plaster-pastedeploy==0.7
    # via pyramid
polib==1.1.1
prometheus-client==0.14.1
    # via cubicweb-prometheus
psycopg2-binary==2.9.3
    # via cubicweb-francearchives (setup.py)
pycryptodomex==3.14.1
    # via cubicweb-francearchives (setup.py)
pyparsing==3.0.9
    # via
    #   packaging
    #   rdflib
pyramid==1.10.5
    # via
    #   cubicweb
    #   cubicweb-francearchives (setup.py)
    #   pyramid-multiauth
    #   pyramid-session-redis
pyramid-multiauth==0.9.0
    # via cubicweb
pyramid-session-redis==1.6.3
    # via cubicweb-francearchives (setup.py)
python-dateutil==2.8.2
    # via
    #   botocore
    #   cubicweb-eac
    #   cubicweb-oaipmh
    #   elasticsearch-dsl
    #   logilab-database
    #   vobject
pytz==2022.1
    # via
    #   babel
    #   cubicweb
    #   cubicweb-oaipmh
pyyaml==6.0
    # via cubicweb-francearchives (setup.py)
rdflib==6.1.1
    # via cubicweb
redis==4.1.4
    # via pyramid-session-redis
repoze-lru==0.7
    # via cubicweb
requests==2.28.1
    # via
    #   cubicweb-francearchives (setup.py)
    #   sickle
rql==0.38.2
    # via cubicweb
s3transfer==0.6.0
    # via boto3
sentry-sdk==1.7.0
    # via
    #   cubicweb-francearchives (setup.py)
    #   cubicweb-sentry
sickle==0.7.0
    # via cubicweb-francearchives (setup.py)
six==1.16.0
    # via
    #   cubicweb-elasticsearch
    #   cubicweb-s3storage
    #   elasticsearch-dsl
    #   glamconv
    #   isodate
    #   logilab-constraint
    #   logilab-mtconverter
    #   pyramid-session-redis
    #   python-dateutil
soupsieve==2.3.1
    # via beautifulsoup4
tqdm==4.64.0
    # via cubicweb-francearchives (setup.py)
translationstring==1.4
    # via pyramid
typing-extensions==4.3.0
    # via logilab-common
urllib3==1.26.10
    # via
    #   botocore
    #   elasticsearch
    #   requests
    #   sentry-sdk
venusian==3.0.0
    # via pyramid
vobject==0.9.6.1
    # via cubicweb-francearchives (setup.py)
waitress==2.1.2
    # via cubicweb
webob==1.8.7
    # via pyramid
wrapt==1.14.1
    # via deprecated
wsgicors==0.7.0
    # via cubicweb
yams==0.48.0
    # via
    #   cubicweb
    #   cubicweb-francearchives (setup.py)
yapps2==2.2.0
    # via logilab-database
yapps2-logilab==2.2.1
    # via rql
zipp==3.8.0
    # via importlib-metadata
zope-deprecation==4.4.0
    # via pyramid
zope-interface==5.4.0
    # via
    #   pyramid
    #   pyramid-session-redis

# The following packages are considered to be unsafe in a requirements file:
# setuptools
